Fix bug where we didn't strip minus of sortedBy argument in sortedBy method of BlogPostRepository,
and thrown an error if we didn't have sortedBy field in BlogPost model. Now we first strip minus, after check for a field in the model. Add few PHPDocs.
Change methods in BlogComment and BlogPost models, which returned describing string if we had a null or false value, now just return null